Replaced interface ModifyAttachmentEventFactory with implementation.
